generate/app/requirements.txt
fastapi
uvicorn
diffusers==0.21.4
transformers==4.37.2
accelerate==0.27.2
huggingface_hub==0.20.3
sentencepiece
rembg
xformers
onnxruntime
python-multipart
Pillow
